requirements.txt
pandas~=1.5.1
python-dotenv~=0.21.0
numpy~=1.23.4
scipy~=1.9.3
implicit~=0.6.1
scikit-learn~=1.1.3
pytest~=7.2.0
fastapi~=0.87.0
pydantic~=1.10.2